pubsub-watermark/src/main/python/requirements.txt
google-cloud-pubsub==0.40.0